raft: make election and heartbeat tick a uint64
In the near future, we want to keep incrementing the electionElapsed
until we hear from a leader. In order to avoid any chances of an integer
overflow, this commit changes the electionElapsed (along with other
symmetrical variables from int to uint64)

Epic: None

Release note: None

+++ pkg/raft/rafttest/interaction_env_handler_tick.go
@@ -31,8 +31,8 @@ func (env *InteractionEnv) handleTickHeartbeat(t *testing.T, d datadriven.TestDa
 }
 
 // Tick the node at the given index the given number of times.
-func (env *InteractionEnv) Tick(idx int, num int) error {
-	for i := 0; i < num; i++ {
+func (env *InteractionEnv) Tick(idx int, num uint64) error {
+	for i := uint64(0); i < num; i++ {
 		env.Nodes[idx].Tick()
 	}
 	return nil
